Finished Server

+++ Server/src/Server.ts
@@ -1,55 +1,45 @@
 import net from 'net'
-import {deserializeStream, deserialize, serialize} from 'bson'
-import {ServerRequest} from './ServerRequest'
+import BSON, {deserializeStream, deserialize, serialize} from 'bson'
 import { EventEmitter } from 'events'
 import * as fs from 'fs/promises'
 import { triggerAsyncId } from 'async_hooks'
-type NetworkServerEvents = "client_login"|"client_leave"|"start"|"shutdown"|"crash";
+import { Packet, Verify } from './Packet'
+type NetworkServerEvents = "client_login"|"client_leave"|"start"|"shutdown"|"crash"|"client_send_data";
 type NetworkServerEvent<T extends NetworkServerEvents> = 
 T extends "shutdown" ? null :
 T extends "crash" ? string : 
-T extends "start" ? null : never
+T extends "start" ? null :
+T extends "client_login" ? number :
+T extends "client_send_data" ? number : never
 interface NetworkServer {
     on<T extends NetworkServerEvents>(eventType:T, event:(event: NetworkServerEvent<T>)=>void):this
 }
 class NetworkServer extends EventEmitter {
     public server:net.Server;
     private clients:{
-        clientName: string
-        sock: net.Socket
-    }[] = [];
-    private clientsPassword:{
-        password:string,
-        clientName:string
-    }[] = [];
+        [key:string]:net.Socket
+    } = {};
+    private clientsIdentification:{
+        [key:string]:string
+    } = {};
     constructor(){
         super()
         this.server = new net.Server()
         this.registerClients()
         this.listenClientsConnection()
-        this.handleErrors()
         this.start(8080, "localhost")
     }
     private async registerClients() {
         const ClientsFolder = await fs.readdir('./Clients')
-         ClientsFolder.forEach(async value => {
-            console.log(value)
+        ClientsFolder.forEach(async value => {
             try {
                 const clientPassword = await fs.readFile(`./Clients/${value}/Password.txt`)
-                this.clientsPassword.forEach(({password}) => {
-                    if(password === clientPassword.toString()){
-                        throw new Error("Password is already been taken");
-                    }
-                })
-                this.clientsPassword.push({
-                    password: clientPassword.toString(),
-                    clientName: value
-                })
+                if(this.clientsIdentification[value]) return;
+                this.clientsIdentification[value] = clientPassword.toString()
             } catch (error) {
-                console.log(error)
-                return
-            }
-        })
+            return
+        }
+    })
     }
     private listenClientsConnection(){
         this.server.on('connection', client => {
@@ -57,20 +47,27 @@ class NetworkServer extends EventEmitter {
                 this.removeClient(client)
             })
             console.log("Client connecting...")
-            client.write("Pls add input password ")
+            client.write("Pls insert Json verification: ")
             client.once('data', async data => {
             try {
-                const clientName = await this.checkIfValidPassword(data.toString())
-                this.clients.push({clientName:clientName,sock:client})
-                this.emit('client_login', client)
-
-                console.log('Client has been successfully logined as ' + clientName)
-                client.write("You have successfully logined as " + clientName)
+                const json = JSON.parse(data.toString())
+                if(!(this.clientsIdentification[json.name] === json.password)) {
+                    client.write("Invalid name or password")
+                    return client.destroy()
+                }
+                if(this.clients[json.name]) {
+                    client.write("A client is already using that profile")
+                    return client.destroy()
+                }
+                this.clients[json.name] = client
+                this.listenSentData(client)                
+                client.write("You have successfully logined")
+                console.log("Client Connected")
                 return true
             } catch (error) {
-                client.write("Invalid password")
+                console.log(error)
                 console.log('Invalid')
-                client.destroy()
+                client.destroy(new Error("Invalid json"))
             }
             })
             client.once('close', err => {
@@ -80,29 +77,37 @@ class NetworkServer extends EventEmitter {
         })
     }
     removeClient(client:net.Socket){
-        this.clients.forEach(({clientName, sock}, index) => {
-            if(sock===client) this.clients.splice(index, 1)
-        })
-    }
-    checkIfValidPassword(input:string): Promise<string>{
-        return new Promise((resolve, reject) => {
-            const max = this.clientsPassword.length
-            let rows = 0
-            this.clientsPassword.forEach(({password, clientName}) => {
-                if(input !== password){
-                    rows++
-                    if(max === rows) reject()
-                    return
-                }
-                resolve(clientName)
-            });
-        })
-    }
-    private handleErrors(){
+        const key = Object.keys(this.clients).find(key => this.clients[key] === client)
+        if(key){
+            delete this.clients[key]
+        }
     }
     start(port: number, host: string){
         this.server.listen(port, host)
         return this
-    }
+    } 
+    private listenSentData(client:net.Socket){
+        client.on('data', data => {
+            try {
+                const json:Packet = JSON.parse(data.toString())
+                if(Verify(json)){
+                    json.sendTo.forEach(clientName => {
+                        const fakeJson:{[key: string]:any} = json
+                        const clientSock = this.clients[clientName]
+                        if(!(clientSock instanceof net.Socket)) {
+                            client.write('Error')
+                            return
+                        }
+                        delete fakeJson.sendTo
+                        fakeJson.from = clientName
+                        clientSock.write(JSON.stringify(fakeJson))
+                    });
+                }
+            } catch (error) {
+                client.write("JSON error")
+            }
+
+        })
+    } 
 }
 new NetworkServer()
